feat: allow materialize with a prefix
allow for materialize subcommand to accept a prefix so all materialized
binaries will be placed under the provided directory.
