+++ cmd/embedded-cluster/materialize.go
@@ -13,6 +13,13 @@ var materializeCommand = &cli.Command{
 	Name:   "materialize",
 	Usage:  "Materialize embedded assets on /var/lib/embedded-cluster",
 	Hidden: true,
+	Flags: []cli.Flag{
+		&cli.StringFlag{
+			Name:  "basedir",
+			Usage: "Base directory to materialize assets",
+			Value: "",
+		},
+	},
 	Before: func(c *cli.Context) error {
 		if os.Getuid() != 0 {
 			return fmt.Errorf("materialize command must be run as root")
@@ -20,7 +27,8 @@ var materializeCommand = &cli.Command{
 		return nil
 	},
 	Action: func(c *cli.Context) error {
-		if err := goods.Materialize(); err != nil {
+		materializer := goods.NewMaterializer(c.String("basedir"))
+		if err := materializer.Materialize(); err != nil {
 			return fmt.Errorf("unable to materialize: %v", err)
 		}
 		return nil
